chore: rename dir model
