+++ src/main/java/com/api/relivro/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.api.relivro.repository;
 
-import com.api.relivro.models.UserModel;
+import com.api.relivro.model.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserModel, Long> {
